Full-text search engine changed to ElasticSearch

+++ TinyShopDb/dbo/StoredProcedures/purchases/spPurchases_Delete.sql
@@ -4,15 +4,18 @@
 AS
 BEGIN
 	SET NOCOUNT ON;
-
-	DECLARE @Quantity INT = (SELECT Quantity FROM Purchases WHERE Id = @PurchaseId)
-
-	DELETE FROM Purchases WHERE Id = @PurchaseId
-
 	DECLARE @CartId INT = (SELECT CartId FROM UserCarts WHERE UserId = @UserId)
 
-	UPDATE Carts
-	SET ItemQuantity = IIF(ItemQuantity - @Quantity >= 0, ItemQuantity - @Quantity, 0)
-	,UpdatedAt = GETUTCDATE()
-	WHERE Id = @CartId
+	IF EXISTS (SELECT Id FROM CartPurchases WHERE CartId = @CartId AND PurchaseId = @PurchaseId)
+		BEGIN
+			DECLARE @Quantity INT = (SELECT Quantity FROM Purchases WHERE Id = @PurchaseId)
+				DELETE FROM Purchases WHERE Id = @PurchaseId
+
+
+
+				UPDATE Carts
+				SET ItemQuantity = IIF(ItemQuantity - @Quantity >= 0, ItemQuantity - @Quantity, 0)
+				,UpdatedAt = GETUTCDATE()
+				WHERE Id = @CartId
+		END
 END
